Added customization.qmd content on project design patterns
